+++ DynamicIsland/ContentView.swift
@@ -1,6 +1,6 @@
 //
 //  ContentView.swift
-//  灵动岛
+//  Dynamic Island
 //
 //  Created by 雪穗子 on 1/7/2026.
 //